Commit more index-* scripts
The existing index-data.sh script runs two indexing operations in the
background.  Have a modified index-single.sh that runs a single indexing
operation in the foreground, making it easier to monitor in a screen
session.  Provide two front-end scripts, index-products.sh and
index-queries.sh, that run index-single.sh with the appropriate
arguments for the two indices that index-data.sh uses.