Implement queries for fetching features
Fill in the code sections to generate queries to be used to fetch the
features that should be used for training a model via XGBoost.

Note that the queries are not currently used; the code that would use
them just fills out random data instead.

It seems sometimes queries return a maximum of 10,000 results, which
appears to be a default limit for OpenSearch.  I suspect there is
probably a problem with sending a query for too many indices.  Attempt
to handle it by breaking queries for large number of indices up into a
list of queries.  The code for using this will need to iterate over the
queries and run all of them, concatenating the results.