Fix heuristic function and downsampling
The code for assigning the heuristic and downsampling it was filling in
random numbers and skipping the downsampling.  Fill in the code to
properly assign a stepped heuristic as requested in the assignment, and
select the proper downsampling function.

This would be much simpler as a few if/then constructs, but I wanted to
try making it capable of handling a much more complicated step function
efficiently.