Remove name_hyphens_min_df from features
The LTR model appears to boost it with max boost, which suggests that it
is not a very useful feature for ranking.

You can see this by running:
  build_ltr.py --analyze --output_dir /workspace/ltr_output --analyze_explains
to generate an explains file.  Inspecting simple_ltr_explains.csv, you
can see that the Feature 2(name_hyphens_min_df) column has some values
over 1e38.  When I did this, I found that there were two queries with
over that value.  One was ``subwoffers'' [sic], and the other was a
concatenation of various SKUs.  I modified the compare_explains() method
so that it dumped the query_obj and the opensearch.explains() for the
``subwoffers'' query, and it showed various scores being boosted by
``maxBoost'' in the LTR results.  It seems XGBoost really cranked the
values there, suggesting this input, for at least some queries, is not
important.  That, plus the nonsensical values it produces, suggests it
should be removed.