Add click_prior_query feature
As described in the assignment, add a click_prior_query feature to the
feature set.  This just copies a parameter that is passed in via the
query, which I already had committed.  In ltr_utils.py,
create_rescore_ltr_query() passes in the click value as a parameter in
the sltr dictionary, which is where the parameter becomes accessible to
the feature set.