Commit answers to questions for other models
Commit my answers to the questiosn posed in the Hello, World Relevance
Edition chapter of the week's assignment.  A previous commit posed the
answers for the simple model only.  Here, I add answers for the hand
tuned model, and the LTR versions of the model, after running
ltr-end-to-end.sh -y (note that I left off -d).

These were supposed to be added to the #questions Slack channel, but
that was inundated with actual questions, so I figured I would commit
them here.