Commit index.py
A script for building an index (and inverted index) to try to figure out
why fasttext reports a connection between the words linux and proteins
in wiki_sample.txt.  Just run index.py while piping wiki_sample.txt into
the input.

Results:  Not too clear.  There are no rows containing linux and
proteins.  There are not even that many other words that occur in both
lines with linux and lines with proteins (aside from words like a, then,
for, and so on).  The closest I saw to a more significant word linking
them together was ``systems'', and that did not occur frequently.