Commit cooking_test
Use a Makefile and a script to automate the commands used to demonstrate
fasttext.

The data directory contains fetch.sh, which was used to download
cooking.stackexchange.tar.gz.  This was then extracted to the cooking
subdirectory, which is used as the source directory in the Makefile.

Add a README mentioning that the commands are available elsewhere in the
repository, too.