Commit generate.sh
A script to go through all of the training, as described in README.txt,
with the options I am trying out for part 3 of this week's course
project.

Update the __init__.py so it reads the correct model location as
generated by generate.sh.

Also, update the logstash file used for week3, so it uses the same Flask
port that has been used in other weeks.