Finish create_labeled_queries.py
Create a CategoryTree class (in utilities/category_tree.py) that stores
Category information in a tree hierarchy (and verified that the input
category information is truly a tree structure in the process).  Have
create_labeled_queries.py load the categories into that Tree, assign
queries to categories, compute the sum of queries in each category
subtree, and then use that to assign each query to the closest ancestor
category with at least the given minimum number of queries.

NOTE:
As a check, this checks that there is only one root node, with the
expected name, and has the expected number of leaf nodes.  The course
work suggests there are 1540 leaf nodes, but I come up with a different
answer (3889).  As described in the comments around the test, I found
out that the 1540 nodes was not counting what I expected it to count.  I
did confirm I could get the 1540 number, but since I had already written
up code to count leaf nodes, I kept the check that ensures it is the
expected value (which I confirmed in other ways should be 3889).  There
are some other tests I tried when I was trying to figure out what
computation might get 1540 as well.