Use query category filtering in search.py
As described in the course materials, modify search.py so it attempts to
map the query to a category.  If the nearest neighbors produce a good
match, those are added to the filters used in the query.  The intent is
to filter the matches to matches which match at least one of the
categories.

Update __init__.py so its default location for loading the query mapping
model matches the output of generate.sh.